+++ Source/Core/ExtentOutlineGeometry.js
@@ -5,6 +5,7 @@ define([
         './BoundingSphere',
         './Cartesian3',
         './Cartographic',
+        './Extent',
         './ComponentDatatype',
         './IndexDatatype',
         './DeveloperError',
@@ -22,6 +23,7 @@ define([
         BoundingSphere,
         Cartesian3,
         Cartographic,
+        Extent,
         ComponentDatatype,
         IndexDatatype,
         DeveloperError,
@@ -302,7 +304,7 @@ define([
         if (!defined(extent)) {
             throw new DeveloperError('extent is required.');
         }
-        extent.validate();
+        Extent.validate(extent);
         if (extent.east < extent.west) {
             throw new DeveloperError('options.extent.east must be greater than options.extent.west');
         }
@@ -344,8 +346,8 @@ define([
 
         var radiiSquared = ellipsoid.radiiSquared;
 
-        extent.getNorthwest(nwCartographic);
-        extent.getCenter(centerCartographic);
+        Extent.getNorthwest(extent, nwCartographic);
+        Extent.getCenter(extent, centerCartographic);
 
         var granYCos = granularityY;
         var granXCos = granularityX;